capstone/requirements.txt
Django~=5.2.3
djangorestframework~=3.16.1
psycopg2-binary
requests
Pillow
celery
redis
